AI/Decision_Tree/requirements.txt
numpy
matplotlib
sklearn
decision-tree-id3